+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [fj-doc-freemarker] generate stub and config conversion now supports registerById and allowDuplicatedId attributes
 - [fj-doc-playground-quarkus] base image registry.access.redhat.com/ubi9/openjdk-21:1.20-2.1729773471
 - [fj-doc-playground-quarkus] added output format ADOC (AsciiDoc)
 - [fj-doc-playground-quarkus] quarkus-version 3.16.1